Factor out repetitive auto code, modify BlueBasket (untested) and create RedBasket (untested); Disable tuning and unneeded op modes in prep for 02/02 competition.
