+++ MeepMeepTesting/src/main/java/com/example/meepmeeptesting/MeepMeepTesting.java
@@ -12,9 +12,9 @@ import com.noahbres.meepmeep.roadrunner.entity.RoadRunnerBotEntity;
 public class MeepMeepTesting {
     static double MAX_VEl = 60,
                 MAX_ACCEL = 60,
-              MAX_ANG_VEL = Math.toRadians(180),
-            MAX_ANG_ACCEL = Math.toRadians(180),
-              TRACK_WIDTH = 15,
+              MAX_ANG_VEL = Math.PI,
+            MAX_ANG_ACCEL = Math.PI,
+              TRACK_WIDTH = 14.2117726773,
                     WIDTH = 17.75,
                    HEIGHT = 16;
 
@@ -33,15 +33,74 @@ public class MeepMeepTesting {
                 .build();
 
         BlueBasket.runAction(BlueBasket.getDrive().actionBuilder(new Pose2d(23-(WIDTH/2), 72-(HEIGHT/2), Math.toRadians(90)))
-                        .setReversed(true)
-                        .splineTo(new Vector2d(55, 55), Math.toRadians(45))
 
-                        .waitSeconds(1)
+                // <preloadScore>
+                .setReversed(true)
+                .splineTo(new Vector2d(54, 54), Math.toRadians(45))
+                // </preloadScore>
 
-                        .setReversed(false)
-                        .turnTo(Math.toRadians(180))
-                        .strafeTo(new Vector2d(-48, 60))
-                        .build());
+                .waitSeconds(1)
+
+                // <outtakeSample>
+                .setReversed(true)
+                .splineTo(new Vector2d(55, 55), Math.toRadians(45))
+                // </outtakeSample>
+
+                .waitSeconds(1)
+
+                // <spikeGrab1>
+                .setReversed(false)
+                .splineTo(new Vector2d(48, 39), Math.toRadians(270))
+                // </spikeGrab1>
+
+                .waitSeconds(1)
+
+                // <spikeScore1>
+                .setReversed(true)
+                .splineTo(new Vector2d(54, 54), Math.toRadians(45))
+                // </spikeScore1>
+
+                .waitSeconds(1)
+
+                // <outtakeSample>
+                .setReversed(true)
+                .splineTo(new Vector2d(55, 55), Math.toRadians(45))
+                // </outtakeSample>
+
+                .waitSeconds(1)
+
+                // <spikeGrab2>
+                .setReversed(false)
+                .splineTo(new Vector2d(58, 39), Math.toRadians(270))
+                // </spikeGrab2>
+
+                .waitSeconds(1)
+
+                // <spikeScore2>
+                .setReversed(true)
+                .splineTo(new Vector2d(54, 54), Math.toRadians(45))
+                // </spikeScore2>
+
+                .waitSeconds(1)
+
+                // <outtakeSample>
+                .setReversed(true)
+                .splineTo(new Vector2d(55, 55), Math.toRadians(45))
+                // </outtakeSample>
+
+                .waitSeconds(1)
+
+                // <trajEnd>
+                .setReversed(false)
+                .splineTo(new Vector2d(48, 24), Math.toRadians(270))
+                .turn(Math.toRadians(180))
+                .setReversed(true)
+                .splineTo(new Vector2d(20, 10), Math.toRadians(180))
+                // </trajEnd>
+
+                .waitSeconds(1)
+
+                .build());
 
         RoadRunnerBotEntity BlueObservation = new DefaultBotBuilder(meepMeep)
                 // Set bot constraints: maxVel, maxAccel, maxAngVel, maxAngAccel, track width
@@ -59,7 +118,75 @@ public class MeepMeepTesting {
                 .setStartPose(new Pose2d(-24+(WIDTH/2), -72+(HEIGHT/2), Math.toRadians(270)))
                 .build();
 
+        RedBasket.runAction(RedBasket.getDrive().actionBuilder(new Pose2d(-24+(WIDTH/2), -72+(HEIGHT/2), Math.toRadians(270)))
 
+                // <preloadScore>
+                .setReversed(true)
+                .splineTo(new Vector2d(-54, -54), Math.toRadians(225))
+                // </preloadScore>
+
+                .waitSeconds(1)
+
+                // <outtakeSample>
+                .setReversed(true)
+                .splineTo(new Vector2d(-55, -55), Math.toRadians(225))
+                // </outtakeSample>
+
+                .waitSeconds(1)
+
+                // <spikeGrab1>
+                .setReversed(false)
+                .splineTo(new Vector2d(-48, -39), Math.toRadians(90))
+                // </spikeGrab1>
+
+                .waitSeconds(1)
+
+                // <spikeScore1>
+                .setReversed(true)
+                .splineTo(new Vector2d(-54, -54), Math.toRadians(225))
+                // </spikeScore1>
+
+                .waitSeconds(1)
+
+                // <outtakeSample>
+                .setReversed(true)
+                .splineTo(new Vector2d(-55, -55), Math.toRadians(225))
+                // </outtakeSample>
+
+                .waitSeconds(1)
+
+                // <spikeGrab2>
+                .setReversed(false)
+                .splineTo(new Vector2d(-58, -39), Math.toRadians(90))
+                // </spikeGrab2>
+
+                .waitSeconds(1)
+
+                // <spikeScore2>
+                .setReversed(true)
+                .splineTo(new Vector2d(-54, -54), Math.toRadians(225))
+                // </spikeScore2>
+
+                .waitSeconds(1)
+
+                // <outtakeSample>
+                .setReversed(true)
+                .splineTo(new Vector2d(-55, -55), Math.toRadians(225))
+                // </outtakeSample>
+
+                .waitSeconds(1)
+
+                // <trajEnd>
+                .setReversed(false)
+                .splineTo(new Vector2d(-48, -24), Math.toRadians(90))
+                .turn(Math.toRadians(180))
+                .setReversed(true)
+                .splineTo(new Vector2d(-20, -10), Math.toRadians(0))
+                // </trajEnd>
+
+                .waitSeconds(1)
+
+                .build());
 
         RoadRunnerBotEntity RedObservation = new DefaultBotBuilder(meepMeep)
                 // Set bot constraints: maxVel, maxAccel, maxAngVel, maxAngAccel, track width
@@ -68,15 +195,13 @@ public class MeepMeepTesting {
                 .setStartPose(new Pose2d(24-(WIDTH/2), -72+(HEIGHT/2), Math.toRadians(270)))
                 .build();
 
-
-
         meepMeep.setBackground(MeepMeep.Background.FIELD_INTO_THE_DEEP_JUICE_DARK)
                 .setDarkMode(false)
                 .setAxesInterval(24)
                 .setBackgroundAlpha(0.95f)
                 .addEntity(BlueBasket)
                 //.addEntity(BlueObservation)
-                //.addEntity(RedBasket)
+                .addEntity(RedBasket)
                 //.addEntity(RedObservation)
                 .start();
     }
