+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/BlueObservation.java
@@ -1,4 +1,22 @@
 package org.firstinspires.ftc.teamcode.auto;
 
-public class BlueObservation {
+import com.acmerobotics.roadrunner.Action;
+import com.acmerobotics.roadrunner.Pose2d;
+import com.acmerobotics.roadrunner.SequentialAction;
+import com.acmerobotics.roadrunner.SleepAction;
+import com.acmerobotics.roadrunner.Vector2d;
+import com.acmerobotics.roadrunner.ftc.Actions;
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
+import org.firstinspires.ftc.teamcode.RoadRunner.MecanumDrive;
+import org.firstinspires.ftc.teamcode.components.*;
+
+@Autonomous
+public class BlueObservation extends LinearOpMode {
+    @Override
+    public void runOpMode() throws InterruptedException {
+        telemetry.addLine("SWITCH MODES > DON'T USE)");
+        telemetry.update();
+    }
 }
