+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tests/ControlHubTest.java
@@ -1,11 +1,13 @@
 package org.firstinspires.ftc.teamcode.tests;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 
 @TeleOp
+@Disabled
 public class ControlHubTest extends LinearOpMode {
 
     DcMotor mPort0, mPort1, mPort2, mPort3;
