+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tests/intakeOuttakePositionTester.java
@@ -2,6 +2,7 @@ package org.firstinspires.ftc.teamcode.tests;
 
 import static com.qualcomm.robotcore.hardware.DcMotor.ZeroPowerBehavior.BRAKE;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -9,7 +10,7 @@ import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-@TeleOp
+@TeleOp(group="test")
 public class intakeOuttakePositionTester extends LinearOpMode {
 
     static int buttonDelay = 60;
